automating-infraestructure-monitoring/requirements.txt
azure-identity==1.16.0
azure-mgmt-compute==31.0.0
azure-mgmt-monitor==6.0.2